+++ Cargo.toml
@@ -42,25 +42,19 @@ metrics-exporter-prometheus = "0.16.0"
 metrics-util = "0.19.0"
 eyre = "0.6.12"
 paste = "1.0.15"
-
-# dev dependencies for integration tests
 parking_lot = "0.12.3"
-lazy_static = { version = "1.5.0" }
 
 [dev-dependencies]
 rand = "0.9.0"
 time = { version = "0.3.36", features = ["macros", "formatting", "parsing"] }
 op-alloy-consensus = "0.12.0"
 alloy-rpc-types-eth = "0.13.0"
-chrono = { version = "0.4.40", features = ["serde"] }
 anyhow = "1.0"
-log = "0.4"
-env_logger = "0.10"
 testcontainers = { version = "0.23.3" }
 assert_cmd = "2.0.10"
 predicates = "3.1.2"
 tokio-util = { version = "0.7.13" }
-nix = "0.15.0"
+# nix = "0.15.0"
 bytes = "1.2"
 reth-rpc-layer = { git = "https://github.com/paradigmxyz/reth.git", rev = "v1.3.7" }
 ctor = "0.4.1"
