+++ Makefile
@@ -41,10 +41,6 @@ lint: ## Run the linters
 test: ## Run the tests for rollup-boost
 	cargo test --verbose --features "$(FEATURES)"
 
-.PHONY: test-integration
-test-integration: ## Run the integration tests for rollup-boost
-	cargo test
-
 .PHONY: lt
 lt: lint test ## Run "lint" and "test"
