+++ RELEASE.md
@@ -18,7 +18,7 @@ Then run the following commands to check the code is working fine:
 ```bash
 make lint
 make test
-make test-integration
+make test
 git status # should show no changes
 
 # Start rollup-boost with the example .env config
