+++ tests/common/test_data/jwt_secret.hex
