+++ tests/common/test_data/p2p_secret.hex
