+++ tests/execution_mode.rs
@@ -1,14 +1,14 @@
+use common::proxy::ProxyHandler;
 use futures::FutureExt as _;
-use integration::proxy::ProxyHandler;
 use rollup_boost::ExecutionMode;
 use serde_json::Value;
 use std::pin::Pin;
 use std::sync::{Arc, Mutex};
 use std::time::Duration;
 
-mod integration;
+mod common;
 
-use crate::integration::RollupBoostTestHarnessBuilder;
+use crate::common::RollupBoostTestHarnessBuilder;
 
 struct CounterHandler {
     counter: Arc<Mutex<u32>>,
