+++ tests/remote_builder_down.rs
@@ -1,10 +1,10 @@
-mod integration;
+mod common;
 
 use std::time::Duration;
 
 use testcontainers::core::client::docker_client_instance;
 
-use crate::integration::RollupBoostTestHarnessBuilder;
+use crate::common::RollupBoostTestHarnessBuilder;
 
 #[tokio::test]
 async fn remote_builder_down() -> eyre::Result<()> {
